Extract Fingerprinter to separate Bazel package.
This is a preparatory step before moving it to the `common-cpp` repo.
