+++ src/main/cc/any_sketch/any_sketch.h
@@ -27,7 +27,7 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/span.h"
 #include "src/main/cc/any_sketch/distributions.h"
-#include "src/main/cc/any_sketch/fingerprinters.h"
+#include "src/main/cc/any_sketch/fingerprinters/fingerprinters.h"
 #include "src/main/cc/any_sketch/value_function.h"
 
 namespace wfa::any_sketch {
