+++ src/main/cc/any_sketch/fingerprinters/BUILD.bazel
@@ -1,0 +1,14 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "fingerprinters",
+    srcs = ["fingerprinters.cc"],
+    hdrs = ["fingerprinters.h"],
+    deps = [
+        "@boringssl//:crypto",
+        "@com_google_absl//absl/base:endian",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:span",
+        "@farmhash",
+    ],
+)
