+++ src/test/cc/any_sketch/BUILD.bazel
@@ -41,14 +41,3 @@ cc_test(
         "@googletest//:gtest_main",
     ],
 )
-
-cc_test(
-    name = "fingerprinters_test",
-    size = "small",
-    srcs = ["fingerprinters_test.cc"],
-    deps = [
-        "//src/main/cc/any_sketch:fingerprinters",
-        "//src/test/cc/testutil:matchers",
-        "@googletest//:gtest_main",
-    ],
-)
